allow disabling achievement notifications #319
This is currently the only configurable notification type that can
only be set to "all" or "none", all the others allow filtering by
sender user, so I had to tweak the drop-down

This opens the way to also eaisly make `note`, `roleAssigned` and
`followRequestAccepted` configurable.

Tested using the API console and sending requests to
`i/claim-achievement`. When the setting was "all", the achievement was
notified and it appeared in `/my/achievements`; when the setting was
"none", the achievement only appeared in `/my/achievements` but no
notification was generated.